hearing_aid: Fix hearing aid behaviour when CoC could not connect
It might happen that CoC cannot be established because of the limited
resorces when HA is connected to other paired device or from any other
reason.
It that happen, Android stays in Connecting state for 30sec which is
Java layer timeout.

With this patch, Java layer is notified about disconnection right after
CoC is not established

sponsor: jpawlowski
test: manual
Change-Id: I682dc77ca6c2b226c69f382058a6cbfc9e082006