Fix a2dp p_db re-use for discovery
A2DP_FindService use a global to store the tSDP_DISCOVERY_DB.
This global was not clean in case of an error process and was reused on
next call. if the size was dynamic (like in the fuzzer) it could
generate an overflow. Current code is using a constant size so the
overflow can not happen. Still fixing this to not be possible

Bug: 203777533
Test: Manual testing to ensure it crash without this fix.
Tag: #stability
Change-Id: Ic170b4a5e1a2bb62dbb229a4db33978ee42cda86