Clear identity bit when passing address to filter
Legacy stack kept Identity Address Type without the identity bit set.
It is later passed without any masking to "Add to accept list", "add to
resolving list", and VSC filter command. All these commands specify that
the identity bit should not be set in related specifications.

This is a CP with partial revert of aosp/2030646.

Test: manual test with DCK application
Bug: 201255908
Bug: 280811285
Change-Id: Id68e81a6706ac40d7394e2803ec3339a7bc44a0a