Have HfpClientConnectionService get updates directly from HFP service
Currently the telecom connection service we have to abstract HFP Client
away to the system listens for braodcast intents to get updates. This
means the updates are blasted to the system, just to end up back in our
own process.

This change makes HfpClientConnectionService a singleton and creates
static event functions for the HeadsetClientService and related classes
to update our connection service with events directly. This will remove
our need for broadcast intents.

Tag: #refactor
Bug: 206035301
Test: atest BluetoothInstrumentationTests
Change-Id: Ib1c02d979cd2320239f904e604e2a1c0917fc489