leaudio: Improve handling audio driver requestes
With this patch we make sure that audio driver will get
ConfirmStreamRequest when ISO stream is in the STREAMING state.
In all other cases, CancelStreamRequest means: please wait.

Also new AudioState::RELEASING has been introduce which is used
for the time when ISO stream is taking down.
When this is happening, we also as audio driver to wait.

Tag: #feature
Test: atest --host  bluetooth_le_audio_client_test
bluetooth_le_audio_test
Sponsor: jpawlowski@
Bug: 150670922

Change-Id: I05cca4c2d6c0ec1f0c2a7f3835d830dd0cc46696