L2CAP: Fix for infinite loop in LE Security Access
Issue:
when there is any pending LE security access queue, the request
will be dequeued from the queue and adding it again as the request
to the same pending queue. So it will be in loop continously for
the same pending request. This loop is not breaking in this scenario.
So other tasks on the BTU thread are not getting processed.

Fix:
All the requests in the encryption pending  queue will processed
one by one. When ever callback needs to send to upper layer, while
loop needs to be break and then execute next pending request after
sending previous security callback to upper layer.

Test: After above steps issue not seen
Tag: #feature
Bug: 208221015

Change-Id: I3e3137ce425b300acc9ba90634dba56e19b3dbc6