Remove config boolean resource for BluetoothInCallService
BluetoothInCallService no longer related only with HFP. In addition,
the config resource value is used in the manifest file and defined as
the initial value of this service. If a bind() request is received from
Telecom while BluetoothService is not started by SystemServer, the
service call flows of BluetoothInCallService such as
onCreate()/onBind()/onUnbind() will be made. In the calling sequence,
the return of getDefaultAdapter(), getAdapterService() become null, and
unexcepted nullpointer exceptions will occur. Since service
enable/disable of BluetoothInCallService is dynamically set by the
packagemanager according to the state transition of AdapterState, the
enable value described in the manifest file is a meaningless value
except when the Bluetooth application is first installed. Moreover,
considering the above case, it seems to be more stable when this value
is false.

Test: Disable BluetoothService and launch dialer with phonenumber
Signed-off-by: Dowon Park <dowon.park@samsung.com>
Change-Id: Ic96ccfcc605d9f965dce43a5b3e446b499b4d69b