LeAudioService: Expose getConnectedGroupLeadDevivces() API
This is going to be used by Telecom.
Each of the group is represented by the Lead devices (simply first
device connected in the group).
This lead devices is used later on when Active device is set in the
Audio System.

Other modules, should be able to know, which of the devices withing
group is a Lead, so they known which one they should make as Active if
needed.

Bug: 150670922
Tag: #feature
Sponsor: jpawlowski@
test: Test: atest BluetoothInstrumentationTests
Change-Id: I3ccf5a20bdd658a647cf1d6837675453224dc90b