Migrating Bluetooth resources to Bluetooth module
In order for the Bluetooth module to build as an apex
all non-public resources must be located in the same
package.
Power usage configs are made system properties to
enable android.internal.os to read them and still have
per device overlays.
Max connected audio devices config is moved to the
Bluetooth module, and getMaxConnectedAudioDevices method
is made system API in order for development Settings to
access it.
All configs used only by the Bluetooth app are moved to
the Bluetooth module.
Vendor overlays will be updated once the migration is done.

Tag: #feature
Bug: 211570675
Test: build, Logging retrieved configs
Test: make RunSettingsRoboTests
Change-Id: Ieed30c31fc44b5b477d43ae120ef11f96ab115ca