Migrating Bluetooth resources to Bluetooth module
In order for the Bluetooth module to build as an apex
all non-public resources must be located in the same
package.
Made getMaxConnectedAudioDevices system API in order
for development Settings to access it.
Removed isInbandRingingSupported as only usage was by bt
service (config is now directly in service).
Removed isBluetoothVoiceDialingEnabled as there were no
usages, and no overlays.
All configs used only by the Bluetooth app are moved to
the Bluetooth module.

Tag: #feature
Bug: 211570675
Test: build
Test: make RunSettingsRoboTests
Change-Id: Ieed30c31fc44b5b477d43ae120ef11f96ab115ca