leaudio: Add method to get device group lead
When Le Audio group is active in the system, it is represented by one
group member which becames Active Device.
With this patch, we want to expose group lead device which will be used
as an Active Device for connected group.

This is needed e.g. by telecom so it knows, which devices should be
presented in the Dialer for each Le Audio group.

Bug: 150670922
Tag: #feature
Sponsor: jpawlowski@
Change-Id: I401ef3ceaeb334b35697dd88e2c1a9b1d5d2eac9