le_audio : Fix issue of LE Audio not connecting after remote disconnect
LE Audio was not getting connected automatically when device in range,
after remote disconnection, VC was getting connected.

This was because the remote device was not added to background connection,
for LE Audio, on remote initiated disconnection, VC was added.

Fix the issue by adding to background connection for LE Audio,
in addition to VC, for non-local initiated background connection.

Test: Verified LE Audio is getting connected after Remote disconnect.
Change-Id: Iac25b1f9cd8cc2d6f29481ebf5abd8fa045a30d7