Fix concurrency issues on state machine shutdown
We should wait for the state machine thread to finish any message
handling as it can still call into the main service. In some cases
the race condition manifested itself when the state machine was
calling okToConnect() on the service when mAdapterService was
already set to null by the main service's stop() call.

Bug: 150670922
Tag: #feature
Test: atest BluetoothInstrumentationTests
Sponsor: jpawlowski@
Change-Id: I75278ccc904881fe20ca7a9451d88ba8720c2740