Cert: Fix cert security accepting any hci events
Some HciMatchers function where returning a function
that return a function instead of a function
that return a boolean. The fact that a function
is a truthy value in python, the matching function
was accepting anything given as input.

By fixing that, this showed an invalid implementation
of CertSecurity.accept_pairing

Test: None
Change-Id: I7f0cfaec6deb0361a3d1e7790b1158c2bcf5ebfd