Pass correct active context from native
When calling LeAudioClientCallbacks.OnAudioConf(), the argument
'avail_cont' was incorrectly being passed. When there is no change
in the active context, the value becomes zero. However it should be
the current active context.

This resulted the LeAudioService make the phone speaker as output
when only one earbud is disconnected, since the service thinks that
the new audio configuration (remaining earbud) is not able to
handle the current audio output.
(See LeAudioService#updateActiveOutDevice())

This CL fixes the available context, and this results:
- Output does not change to speaker when only one headset is
  disconnected (i.e. another side of headset is not disconnected yet).
- Then the output changes to speaker when the remaining headset is
  disconnected. Music is paused at this moment.

Bug: 207090924
Test: Maunally tested followings:
      - forgetting LE headset (music paused)
      - disconnecting LE headset (music paused)
      - Switching btw. LE headset and speaker via UI (music not paused)
Change-Id: I916d6dd044b15b477700d27db8f25025f2404f4f