floss: Add call status and hook to CLCC callback
The report of the call status indicator is required for some headset to
work properly. Add a call_status private field and abstract the set of
the call status to a private helper function for better readability in
the ConnectionStateCallback. Also hook the CLCC callback to reflect the
correct call status.

Tag: #floss
Bug: 213408429
Test: Build, flush and verify the btsnoop
BYPASS_LONG_LINES_REASON: Bluetooth likes 120 char lines

Change-Id: Ib9af4bf0ebe695ede4614703a6d0ca2adf5478a8