[DO NOT MERGE] le_audio_test_app: Legacy BASS code for reference
Note that this is a legacy code and does not compile against the existing
Broadcast Scan Assistant API. This code was previously removed from the app,
and after that the app was refactored so this would probably not compile
even if the API is the same. This patch is for reference only.

Change-Id: Ib13e9e61ac4fdba31102889aa0efd4a0717cff54