Stay on the IDLE mode after the reconfiguration
BT stack should not auto start stream after the reconfiguration. The
stack may not if there is audio data need to be streamed. The example is
as phone call ended, and no media need to be resumed, the stack would
always keep on streaming state unital audio HAL and audio framework
trigger startStream again.

Bug: 222674521
Test: Make build
Change-Id: I548c4b91e41c1f3e9112adeb9c780045278fd412