Added Bluetooth audio device fallback on disconnect
When disconnecting either the Hfp or A2dp active Bluetooth
device, ActiveDeviceManager will now check if there are
other connected devices and set active the most recently
connected one instead of routing the audio on the phone.

This will apply only when a disconnection happen on the
active device, other cases such as Hearing aid connection,
or wired connection are not impacted by this change.

In summary, if the active Bluetooth audio device is
disconnected and other audio devices are connected, the
audio will be routed to the most recently connected device
instead of the phone.

In the case of A2dp, the music will not be paused when
switching from one device to another, but will still be
paused when the last Bluetooth audio device has been
disconnected.

Bug: 202602952
Test: manual
Tag: #feature
Change-Id: Ib935093bc2729500ce93d1a8f80dcf8f310f263b