Added Bluetooth audio device fallback on disconnect
HeadsetService#removeActiveDevice and A2dpService#
removeActiveDevice will now check if the device being
removed is not connected anymore. If so, and if another
device is connected to these profiles it will set the other
connected device active.

This will apply only when a disconnection happen on the
active device, other cases such as Hearing aid connection,
or wired connection are not impacted by this change.

In the case of A2dp, the music will not be paused when
switching from one device to another, but will still be
paused in other cases.

Bug: 202602952
Test: manual
Tag: #feature
Change-Id: Ia77d3d0520ff5f65779c9ea15ba0c1aa7eb1476b