Avoid crash on unexpected command status/complete
For test need, tester often has need to send a HCI command to
controller without stack intervention (e.g by doing an echo write
to Bluetooth controller's UART port on adb shell). It may be
enough for the stack to warn (than crash) and ignore the command
status or command complete event for this case.

Test: Verified stack does not crash on receiving unexpected
      command complete

Change-Id: I845bbef0da25a6943361989149d5474ae5cab95e