Avoid crash on unexpected HCI command status/complete
For test purposes, need to allow sending HCI command to
the device without stack intervention (e.g an echo write
of command string to BT UART on adb shell). This change
softens BT stack approach on receiving an unexpected
command status or command complete event that occurs as
a consequence of the above said action. Change is to only
warn (than crash) and ignore those events.

Test: manual

Change-Id: I845bbef0da25a6943361989149d5474ae5cab95e