pdl: fix computation of `SourceLocation` line numbers
Before, we would always exit the loop at the first iteration since the
`line_starts` passed in start with a zero offset. We should actually
search through the line starts to find the last line which starts
before the given offset.

This also adds new unit tests to cover this.

Test: atest pdl_inline_tests
Change-Id: I171803822c3cb58deec167a2bc063cbd8f99110f