PDL: Unify the declaration scope for Packet and Struct declarations
Packet and Struct currently live in different scopes, which allows
packet and struct to have the same name. This impacts the rust
code generation as packets and structs are disambiguated by adding
appropriate suffixes, which is undesirable.

This change renames the few name conflicts in hci_packets.pdl that
prevent an immediate enforcement of this new rule.

Bug: 228327522
Test: Tested libbluetooth, root-canal target compilation
Change-Id: I88bbd394b56dbf6ac24ead4dcdeddcd467f71426