PDL: add basic support for generating Rust code
This CL adds support for handling a simple PDL file which defines a
packet. The CL also includes infrastructure for testing the generated
code against known-good output. This is sometimes called “snapshot
testing”. While there are crates for this, we currently don’t use any
to avoid pulling in new dependencies.

Bug: 228306440
Test: atest --host pdl_inline_tests
Change-Id: If33403444ba9e4ce9c45e15b65c1da03011aa5e9