Floss: Fixing classic keyboard SDP and connection
If service discovery is ongoing, the connect_all_enabled_profiles()
might not have any profile to connect. In that case, we will retry
again once the UUIDs from a remote device are available.
Besides, we also remove the cancel_discovery() from
connect_all_enabled_profiles() as it could interrupt the service discovery
once a device is bonded.

Bug: 219021367
Tag: #floss
Test: manual test on BR/EDR keyboards and mouse
Change-Id: Ie0445520cfed9b5fa821a08f116df16ea6e45073