BlueberryServer: Always play audio track
In some runs of PtsBot cases a2dp is already playing
(`isA2dpPlaying` returns true) when we start. In this
case we don't call AudioTrack#play making the playback
AudioTrack#write calls fails.
We should investigate why a2dp is playing instead but
in the meantime this is a simple and effective fix.

Test: atest pts-bot
Change-Id: I67d81723fb0c468f8629e0bfe4c43adac59a709c