Broadcaster: Fix not notifying about the started broadcast
The existing code expects the initial broadcast state to be
STREAMING. This turns out to be wrong as the native stack currently
sends the initial state as 'PAUSED' and expects the service is to
actually trigger sending the broadcast audio stream as an additional
step.

This change fixes this by calling onBroadcastStarted() right after
the broadcast is created and then notify any playback state changes
when the native code switches to STREAMING.

Unit tests for broadcast starting and stopping are extended with
callback call verifications.

Bug: 150670922
Bug: 229695142
Fixes: 229695142
Tag: #feture
Test: atest BluetoothInstrumentationTests
Sponsor: jpawlowski@

Change-Id: I0f4618f636f22c09c5abc11373218c753fe46177