leaudio: Improve cleanup on Bluetooth stack off
Cleanup() is called when Bluetooth stack is off via GUI.
However, Bluetooth chip still might work for other purposes.
Therefore on Cleanup we make sure that all CISes and ACLs are
disconnected.
Note: In case of Bluetooth being Off while, streaming, we might end up
with a CIG created in the controller. It will be cleared on Bluetooth On
event, as HCI Reset is sent to the controller.

Bug: 150670922
Bug: 228351333
Test: atest bluetooth_le_audio_client_test
Change-Id: Ia5ad66b0949ed693499efc88de7ffc783508a4c3