leaudio: Configure only connected device
In futher work, we will be configuring also not connected devices to be
able to prepare CIG for the whole group even one device is connected.
But for the moment, our assumption is that we configure only connected
devices. Without this patch, we could incorrectly send Codec Config to not
connected device when playing around with earbuds beeing disconnected
and reconnected.

Bug: 150670922
Bug: 228809876
Sponsor: @japawlowski
Test: atest --host bluetooth_le_audio_test
bluetooth_le_audio_client_test

Change-Id: I401b1890ac1c5f550de052b77e355f5679c460cf