WIP: floss: Integrate Eiffel SBC library
This is a WIP CL to integrate the Eiffel SBC library into Floss. We need
to do this development behind a flag so we don't affect the existing
code in Fluoride and have a plan for the eventual switchover/default
usage of Eiffel SBC.

See go/floss-sbc-lib for more info.

This code is incomplete and needs the following to work:
* git clone <eiffel sbc location> system/embdrv/sbc/esbc
* rm -rf system/embdrv/sbc/esbc/.git
* Implement the missing functions for a2dp_sbc
* Integrate this into the SCO recv/send paths to do mSBC encoding

Bug: None
Tag: #floss
Test: None
Change-Id: Ic801ecea2e0a692aa54674c9b89dc8de849e901d