floss: Qualify moves as std::move
Originally, these moves all were preceeded by std::.
However, this was changed a few years ago to improve the
signal-to-noise ratio when reading the code to omit the
std:: through using std::move.

However, this isn't recommended by upstream clang at all,
move in particular should always be preceeded by the appropriate
qualified std:: namespace. This warning is being enabled
in newer versions of clang, and is a fatal warning for the
ChromeOS builders.

This change is required for LLVM roll r458507 in CrOS.

Bug: 231351802
Test: CQ
Test: emerge-amd64-generic net-wireless/floss # with llvm-next
Change-Id: I5191ba8b4a69f4ebc745dd4c7122c218b7b5d9da