floss: Remove noop std::move(task_)
Moves, in general, are semantically equivalent to a static
cast to an xvalue. Unless there's some std::move overrides
here that really shouldn't happen, this std::move(task_) is
actually a noop.

This is currently giving a -Wunused-value compiler error,
and needs to be updated for the new LLVM roll to r458507.

This change is required for LLVM roll r458507 in ChromeOS.

Bug: 231351802
Test: CQ
Test: emerge-amd64-generic net-wireless/floss # with llvm-next

Change-Id: I893217cf0966a7a62e30f4bd2c0a46791e9f016b