floss: Always use newavrcp path in system
A recent sysprops refactor changed the default value of a sysprop from
true to false and this caused a legacy AVRCP path to be taken which is
not well tested. Update the check so that the new avrcp path is always
taken and doesn't depend on the sysprop.

Bug: 230474915
Tag: #floss
Test: ChromeOS automated tests

Change-Id: Ifea8b071c6dbffb2978a5139ea7b304082d2cf34