Floss: Workaround D-Bus timing
Under D-Bus spec, when a service announces an availability of an
Interface via InterfacesAdded signal, it signifies that it is ready to
accept calls on that interface. However, it seems that `dbus-rs` library
does not guarantee the readiness of accepting calls after emitting
InterfacesAdded.

This is a temporary workaround to wait for a certain amount before
making method calls after receiving InterfacesAdded of btadapterd's
Suspend API.

Bug: 224606285
Tag: #floss
Test: Manual - Enabling adapter should not crash btmanagerd

Change-Id: Iad6a932e6d4cf99dc6857b015052405f1879c568