pdl: enable snapshot testing via ‘cargo test’
This adds a minimal infrastructure needed to support snapshot testing
with Cargo. In short, we read a snapshot file with known-good content
and compare this to the generated output. If they differ, we
optionally override the snapshot with the generated output. This will
allow us to easily update the output as we evolve the code generator.

This the strategy described in go/pdl-rust-backend-testing-strategy.

Test: atest pdl_inline_tests && cargo test
Change-Id: I82078f77168ee7141de71ff57935c1fb12d09364