LeAudioService: Provide Ccid to the native code
This patch extends ContentControlIdKeeper with more information about
the user of the ccid e.g. Context type and userUuid.
The userUuid should be a GATT service uuid which used Ccid or any other
uuid. Note that LeAudioService filters ccid with GTBS and GMCS uuid.

This patch also introduce onInitialized() callback from native to java.
This is needed for LeAudioService to find a good moment to feed Native
with known Ccids

Bug: 230551601
Test: atest BluetoohtInstrumentationTests
Test: connect audio and verify logs in ASE Enable command

Change-Id: I55d8674c86575ba8a91d9c2a1885d02e1009b919