Floss: Improve D-Bus projection for client
This extends the client D-Bus projection generation to also generate the
more RPC-friendly struct:
* Methods are async allowing clients to yield instead of block. Although
  all Floss methods are sync at the server side, at RPC-level it still
  has benefit to yield rather than block for the return.
* Methods have returns wrapped in Result<T, dbus::Error> allowing
  clients to handle D-Bus-level errors.

Included in this patch are a few examples on how the RPC-friendly
version can be used by btclient, but these examples do not actually
fully handle the error yet. Future patches will be able to utilize this
RPC-friendly interface to fully handle D-Bus errors.

Bug: 200732080
Tag: #floss
Test: Manual - Ran btclient without btmanagerd

Change-Id: Ia373fa88354613b3a2ce04b4711b453c099ac14e