Block contact download until we can use accounts
The account manager framework has a race condition in it where theres as
much as a 2 second delay between when our service is enabled and when
the framework recognizes our service and actually lets us create and
destroy accounts.

We previous enabled the service by default in the manifest, but they
caused remnants of it to appear on devices that don't support PBAP
Client. Now, we disable it by default, but block usage of accounts until
we know they're available to us. We wait as much as 3 seconds, checking
every 500ms to see if our accounts are visible. If we don't get access
to our accounts in those 3 seconds, we log an error.

The service being off when PBAP Client is off will cause the remnants
of the service to disappear on devices that don't expect it.

Tag: #stability
Bug: 229629444
Bug: 232943969
Test: atest BluetoothInstrumentationTests
Merged-In: I8d57056dd1ef2b017557ef9d0d3fbbae24e553fa
Change-Id: I8d57056dd1ef2b017557ef9d0d3fbbae24e553fa