floss: Refactor insertion via dbus projection
Dbus projection currently requires that a single dbus object path only
has a single interface associated with it. This results in multiple
objects being created for several interfaces.

This change refactors generate_dbus_exporter to provide an Interface
token which must be separately inserted with the object path. It also
provides a mechanism to have Mixins, a single object that contains
multiple interface implementations.

Bug: 233123489
Tag: #floss
Test: ./build.py and manual test on ChromeOS
Change-Id: I82f2cd18d02276bf26f77541ba3592a73ca558fa