Check context type with AUDIO_USAGE_VOICE_COMMUNICATION_SIGNALLING uasge
As dail pad tone sound, the usage would be AUDIO_USAGE_VOICE_COMMUNICATION_SIGNALLING with context type unknown.
When bt stack recieve this usage metadata during media streaming, it would trigger the reconfiguration to conversational, because the conversational is higher priority than media in the general case.
So we check context type with
AUDIO_USAGE_VOICE_COMMUNICATION_SIGNALLING as well to avoid the
unexpected reconfiguration

Bug: 231411704
Test: during media stream, and make the dial pad tone simutaously, make
sure the media
streaming normal

Change-Id: If14658e94282644d837bedff089e4029c7fc334c