Expose HciTransport in test environment
With this change we expose the HciTransport interface to external users.
This will enable:

- The gRPC endpoint to directly forward hci packets
- Exposition of the hci_sniffer to external users

This introduces an AsyncChannelServer that notifies listeners when
a connection of type T is produced. The server can be used to produce:

asyncDataChannel or a rootcanal::HciTransport. An adapter class is
introduced that turns raw sockets into an hci transport.

Bug: 231598376
Tag: #refactor
Test: Two emulators can discover each other
Test: gd/cert/run
Change-Id: I2426459538a1d55ed685eeeb2c4241e3996c091c