Add tool for plumbing BT API interfaces.
This tool will plumb APIs with the given names from the
Topshim layer to the GD Controller layer.

 Topshim
  |
  v
 BTIF
  |
  v
 BTA
  |
  v
 BTM
  |
  v
 Controller Shim
  |
  v
 GD Controller

It doesn't account for arguments, you will need to manually add them.

It doesn't account for inband callbacks either, you will need to manuall add
those.

In some cases your code may not compile immediately, but it is usually
only 1 or 2 syntax errors in the controller shim/gd controller layer if
it happens at all.

The script will auto commit and rebase for comment editing to add things
like the following...

Bug: 235505882
Test: mma -j $(nproc)
Test: ./build.py
Change-Id: Iabe05192b8adbd0067d9146139f6b4a6efbdf20a