Do not wtf when stack sends event to stopped headset service
The remote device initializes the connection during Bluetooth turniing
off process, the stack receives the AT commands and sends to the
HeadsetService, but it is already stopped. A crash is triggered in
this kind of scenario. Since this scenario won't impact functionality,
we may just print error message to instead of crash.

Bug: 212644286
Bug: 235764096
Test: BCST test
Change-Id: I8c31ea4e337e9851c3410f34e9167e25e81e0775