Remove rfcomm_security_records map, store records in tBTA_JV_PCB instead.
The map is globally keyed by SCN, which causes issues when SCNs are
reused (server vs client, or different device connections). tBTA_JV_PCBs
are distinct for every (port, device, server/client) tuple, and so fit
this use-case.

I will write a test using the Avatar framework shortly to verify the fix, but for the time being this will be tested manually

Ignore-AOSP-First: Bugfix for S
Tag: #stability
Bug: 230553684
Test: Manual, automated test in followup
Change-Id: Iae24f07fd5d19f66126abfe61052fab96d1b1bdf