Pandora: Implement RFCOMM/
RFCOMM/DEVA/
RFCOMM/DEVB/
RFCOMM/DEVA-DEVB/

There are two tests that are still skipped.

I couldn't find the IXIT value referenced in the test spec,
which would allow the tests to pass:

The role of DevA/DevB taken on by the IUT does not
matter to achieve a pass verdict for some test cases
for this Protocol as indicated in the test purpose and is
specified in the test case identifier for role agnostic
tests as detailed in Table 4.1. The role of the IUT/Lower
Tester for these tests is specified in the IXIT [7] in
order to enable the correct test environment conditions
to provoke the Lower Tester.

Since we don't support hooking up a physical serial_port,
TSPC_SPP_2_1 is set to false, which disables
RFCOMM/DEVA-DEVB/RFC/BV-14-C

Bug: 240483035
Test: atest pts-bot:RFCOMM/ -v
Change-Id: I7df1d38f1a5501a42030d6dfeff5560eb1124fe2