SDP: Set address in tBTM_REMOTE_DEV_NAME
The tBTM_REMOTE_DEV_NAME.bd_addr was never set but we do use it to
decide if we can remove the callback in bta_dm_remname_cback.
Without the fix we are triggering too many remote name request. They
will eventually conflict with each other and ends in the busy state.
At this point if the stack is performing a search, we could receive
delayed sdp_result that will end in a crash.

Repro steps:
 * Start a search on phone
 * Pair & connect from carkit
 * Crash should occur (1~4% repro).

Without any logs nor fix, crash is around 15%+ repro
With the fix, the 'crash situation' only appear around 1~2% and it no
longer crash
Crash situation as been identified to match with the following logs
    `BTM_ReadRemoteDeviceName is busy`

Bug: 231150390
Test: Execute repro step above
Merged-In: I317f5d689a9efb85d997a38470712ed3d8e0c531
Change-Id: I317f5d689a9efb85d997a38470712ed3d8e0c531