Fix admin enable/disable of Bluetooth file sharing, part 2
This fixes the bluetooth package detection functionality in the case
when *all* components are disabled. This occurs in the CTS test, but did
not occur in manual testing of ag/19219300.

Also did manual testing to make sure we don't regress (toggled the no_bluetooth_sharing option, verified that it appeared/disappeared)

Bug: 239372439
Test: atest CtsDevicePolicyManagerTestCases:DeviceOwnerTest
Tag: #stability
Ignore-AOSP-First: depends on a ton of changes that are only in tm-dev
Change-Id: Ib9530a8829a946ea00a21111e459d697086f2ef8