btif: fix bond state callback state mismatch
When bonding is not finished successfully with authentication, btif_dm
should report a state update to the Java layer if: 1. the device is
still bonding or 2. the device has been removed. However, the existing
code always reports the BT_BOND_STATE_NONE state; fix it by separating
the handles for the two situations.

Bug: 232098267
Test: ./build.py
Change-Id: I8a7e73f3fc3f447d84f45b97493d0556c493d74c