Do not cancel remote name request during pairing.
When the user taps on a device to pair, we cancel inquiry and proceed
with creating a bond with the remote device. As part of bonding, we send
a remote name request to the remote device. However, the inquiry
cancellation that we triggered earlier ends up cancelling this remote
name request. Since the remote name request did not complete
successfully, pairing fails.

In the code, the rationale for sending a remote name request seems to be
that a positive response would mean the remote device supports BR/EDR
2.1+ which means security mode 4 can be used. Since this remote name
  request is cancelled, we treat it as remote device not supporting
  security mode 4 and thus fail pairing.

Bug: 238914503
Tag: #refactor
Test: Manually paired several times, unable to repro the issue

Change-Id: Ia3674780d129920c5b69551c34ab836c65e6d4d2