Refactor remote name lookup to use queuing
This CL centralizes all the remote name lookup logic. We have a queue that keeps track of pending requests, as well as a static variable tracking the active request. An opaque handle is returned to the caller when they start a request, and is used to indicate which request needs to be cancelled. Callbacks are only invoked when their corresponding request returns, so they don't need to handle any retry/address verification logic. If we cancel a request that has not yet been dequeued, we remove it from the queue and synchronously invoke the callback. If multiple requests are made for the same remote name, only a single request is actually sent, but each request gets their own handle (for cancellation).

Previously, we had an ad-hoc way of sending out remote name requests and registering callbacks. They could be sent from either the SEC module or the DM module. The previous logic is summarized below:
```
SEC requests go out NO MATTER WHAT
Only one DM request can be outstanding at a time.
  If one DM request is outstanding, we add bta_dm_service_search_remname_cback to the SEC callbacks.
  Otherwise, we set bta_dm_remname_cback as the DM callback.

When any request completes, we invoke all SEC callbacks as well as the DM callback, if available.
We also invoke btm_sec_rmt_name_request_complete, which (I THINK) only does SEC related stuff.

When bta_dm_remname_cback runs, if the request addr matches the DM addr, we remove the DM callback
from the SEC list. In addition, *no matter what*, we report REMT_NAME_EVT back to the state machine.

When bta_dm_service_search_remname_cback runs, we check if the request addr matches the DM addr.
If it does, we succeed and invoke bta_dm_remname_cback. Otherwise, we re-propose the request.
If the proposal is accepted (either directly or via a SEC callback) we exit. Otherwise, we report
failure via bta_dm_remname_cback.
```

This is buggy, since we're always triggering the btm_sec_rmt_name_request_complete callback, and can potentially have multiple requests in-flight at once. When we re-propose the request, depending on the current state, we could potentially end up registering multiple callbacks at once. In addition, if a request fails and passes into `bta_dm_remname_cback`, we report the REMT_NAME_EVT twice, breaking the DM state machine. Request cancellation also just causes all the outstanding requests to be cancelled, leading to pairing failure (if the SEC module is waiting on a remname request and the DM module cancels it, we get a spurious pairing failure).

Change-Id: I3810bad743930d53264d73971c172b741428eb31