Refactor remote name lookup to use queuing
This CL centralizes all the remote name lookup logic. We have a queue that keeps track of pending requests, as well as a static variable tracking the active request. An opaque handle is returned to the caller when they start a request, and is used to indicate which request needs to be cancelled. Callbacks are only invoked when their corresponding request returns, so they don't need to handle any retry/address verification logic. If we cancel a request that has not yet been dequeued, we remove it from the queue and synchronously invoke the callback. If multiple requests are made for the same remote name, only a single request is actually sent, but each request gets their own handle (for cancellation).

Test: TBD
Bug: 240444535
Tag: #refactor

Change-Id: I3810bad743930d53264d73971c172b741428eb31