Minimal fix for SDP crash
When we invoke bta_dm_remname_cback, if we are waiting for a response
from a request sent in the DM module, but instead got a response from
SEC, we want to ignore it entirely, rather than updating the state
machine but then continuing to accept other events.

This is an awful hack on top of lots of awful hacks - ag/19412615 will fix this properly, but it's too large of a change to release for T at this point.

Bug: 231150390
Tag: #stability
Test: manual
Change-Id: I416fd112c270cc5725aec378cc9698a10911f02e