Added list instance removal in GattInstance#disconnect()
PTS is trying to connect a device that has not been fully
disconnected, but won't accept that the grpc waits for the
diconnection before returning. We have to remove the
instance as soon as we call disconnect, but we also have
to remove it when the disconnection comes from the device.

Bug: 239103399
Tag: #feature
Test: atest pts-bot
Change-Id: I808f07d1434fa160ab057528e4e21313348505fe