Floss: extend internal state wbs_supported for multi-device
Currently CRAS only supports single audio device, thus the internal
states may be shared.

However, wbs_supported is unique in a way that it
could be hung on stale state when an NBS-only HF device is connected
after a WBS-supported HF device was, in which case no codec-negotiation
would happen to update it.

In this CL, we use a dictionary to store wbs_supported, and introduce a
reset function to prevent such stale state in the future.

Bug: 241014631
Tag: #floss
Test: Build and verify CRAS correctly reads wbs_supported
Change-Id: I95e76895c4974149bb97d03c90957ba293e1f169