Remove dependency PlatformProperies from Bluetooth app
packages/modules/Bluetooth/android/app/Android.bp include
PlatformProperties as dependency as well as
"framework-bluetooth-pre-jarjar". The last one also adds the
dependency PlatformProperties, which it is causing the test
StrictJavaPackagesTest.testApkInApex_nonClasspathClasses to
fail.

Note:

The problem the test is trying to detect is multiple copies
of the same class across the bootclasspath; in this case,
framework-bluetooth is part of the bootclasspath, so all its
classes are by default available (with caveats) to all java code;
PlatformProperties is being bundle again in the Bluetooth APK
(and applying the same jarjar rules as framework-bluetooth),
so the end result is that we are bundling the same class that
is already in the bootclasspath.

Bug: 231718727
Test: Run packages/modules/Bluetooth test cases.

Change-Id: Ia7015fe1dcd279b5e39853a234fef7d908eb8236