Configure build flags depending on host machine
When compiling for Raspberry Pi, various flags were required to get the
build to work, such as flags setting the CPU architecture. This change
allows for adding C/C++/Rust flags to the build depending on the host
machine.

Bug: 241967798
Tag: #floss
Test: run ./build.py on an armv7l and note that correct flags are being
adding to the build, and note that these same flags are not added to a
non armv7l build.

Change-Id: I3937dc694cf720bbb17f501f2d2252b5833ccc02