Save UUIDs using properties rather than the pairing_cb_t
The pairing_cb_t can be reset due to pairing state changes. We currently
store UUIDs in the pairing_cb_t structure. Instead of that, we should
save them using properties so that we can retrieve them later even if
pairing_cb_t was reset.

1. Upon discovery of classic UUIDs, get the existing "interesting" BLE
UUIDs that we know of (if any) and union them with the newly discovered
classic UUIDs.

2. Upon discovery of BLE UUIDs, get the existing UUIDs that we know of
   and union them with newly discovered "interesting" BLE UUIDs

This incorporates suggestions from Samsung because the original change
was causing issues with LE Audio.

Bug: 235407715
Bug: 239483439
Bug: 242007275
Tag: #refactor
Test: Samsung was running into issues with LE Audio and suggested the
change after they tested it. I tested it manually using a JBL Flip 5.
Change-Id: I31101f2607a3fe63bfec3b482f4dafceebe0ebe4

Change-Id: I9eea05b7338dc05b17f54778e2d1189ec6729ebf