floss: Move the read buffer out of btm_route_sco_data
We change the behavior to not drop PCM data we read when its length is
less than BTM_MSBC_CODE_SIZE bytes by moving the buffer out of the
function.

The new behaviors after this CL will be:
1. We don't drop PCM now if audio server send PCM less for one mSBC
   packet
2. Do not send zero frame to SCO if not enough

With these new behaviors, we'd like to make the audio frames eventually
played continual and without pop noises, caused by the added zero frames.

(1) is what we want to improve in this CL compared with existing
behavior. We should take extra attention for (2). (2) means that we
intentionally let SCO Rx/Tx offset by 1 frame when audio server underrun.
Since we don't add any mechanism to catch up. It is fully relies on the
controller to handle potential underrun in SCO.

This is different from how we implemented CRAS (ChromeOS Audio Server)
for BlueZ that we fill zeros when we have less than a mSBC packet amount
of PCM data to write. However, we still think it is better to do so as
adding mechanisms to handle the underrun on the Bluetooth will:
- Introduce clear pop noises between the valid audio frames frames
- Add the complexity for audio stack to coordinate with the Bluetooth
  stack with these extra implicit behaviors.

Bug: 235901463
Tag: #floss
Test: gd/cert/run & Build and verify HFP works & cras_test_client \
--block_size 512 --playback_file 400.wav

Change-Id: I8323fdd7eb3af899722fd1aa3b6643d3f9a64bb6