floss: Move the read buffer out of btm_route_sco_data
Moving the buffer out of the function enables us to change the behavior
that we always drop PCM data we read when its length is less than
BTM_MSBC_CODE_SIZE bytes.

The new behaviors after this CL will be:

 1. We don't drop PCM now if audio server send PCM less for one mSBC packet
 2. We fill in zeros to PCM buffer to balance the SCO Rx/Tx rate to
    secure that we send one corresponding Tx SCO packet for every Rx
    packet we received.

(1) is what we want to improve in this CL compared with existing behavior.
We should take extra attention for (2). With (2), we will fill zeros to
mitigate the scenario that the audio server has less than requested frames
(underrun).

We believe this behavior should be safe as it is actually close to how we
had implemented CRAS (ChromeOS Audio Server) for BlueZ that we always
fill in zeros. For the implementation, check:
https://source.chromium.org/chromiumos/chromiumos/codesearch/+/main:src/third_party/adhd/cras/src/server/cras_sco.c;l=315;drc=5197d72f5ee79678f8effa56da9db4c8f48c809c

Bug: 235901463
Tag: #floss
Test: system/gd/cert/run & Build and verify HFP works

Change-Id: I8323fdd7eb3af899722fd1aa3b6643d3f9a64bb6