Use media configuration as default to avoid the reconfiguration
We use Media as the default context type. List some audio
configurations for the specific context usage. But for the other context
type which doesn't have the audio configuration list, we will use the
default configuration. It would trigger the reconfiguration and have
some delay as context switch.

Some the patch aline the default configuration with the media one, it
could reduce this kind of delay. Just change the context type, but don't
do the reconfiguration

Bug: 236667384
Test: Start the navigation as no media streaming, check the single
behavior

Change-Id: I524a59b56a1056335e80b1b0c73224e7e56d658d