Floss: fix wbs_supported for sco_hfp_hal_linux
Currently the preferred codec is overriden to mSBC even when the local
device does not support WBS, mainly due to how Android expects WBS to
always be supported (and can be decided in compile-time if not).

In this CL we ignore the "AT+BAC" event that attempts to modify the
preferred codec to WBS when local device does not support it
(hal-independent). Also, we guard `get_wbs_supported` in
`sco_hfp_hal_linux` to align with that of Android and for better
testing.

Bug: 243732454
Tag: #floss
Test: Build and verify DISABLE_WBS is reflected to `wbs_supported`
Change-Id: I06e3bf905ccb75f4d86b035f7da04ff7f5737547