[PANDORA] Store intent to be used by delayed callback
Pts is trying to connect to device and then ask to accept the popup.
If we start listening on the intent too late we may miss it.
Here we are storing all the usefull intent in a temporary list to
retrieve it from. The lifetime of such intent is 100ms, after that we
will not allowed them to be re-used.

Test: atest pts-bot
Fix: 242326310
Change-Id: If71eb0e396c50a8f217a41d56b35030c7fdac4f5