floss: Fix AvrcpAbsoluteVolumeChange race
We'll wait for a period of time for both A2DP and HFP's capability to be
updated before sending the device added event. If
AvrcpAbsoluteVolumeChange is triggered when we are waiting, the audio
stack will regard it as a abnormal event as the device is not added on
its point of view. The following device added event also won't help as
the absolute_volume value in that event is out-dated.

Cancel the pending event and trigger a new one with the correct
absolute_volume information to solve the race.

Bug: 243083916
Tag: #floss
Test: Build and verify volume control works
BYPASS_LONG_LINES_REASON: Bluetooth likes 120 char lines

Change-Id: Ib76651a77fc3116c2d5b90dbabffe98dfb27424e